Добавляет контроллер доски
